feat: Refactor TabMenu components and updated HealthData page
This commit refactors the TabMenu components by introducing the NavTab and NavTabs components. The NavTab component is a styled Tab component that adds custom styles for selected and unselected states. The NavTabs component is a styled Tabs component that removes the default indicator. These changes improve the visual consistency and user experience of the TabMenu. Additionally, the HealthData page is updated to use the new TabMenu components for displaying different sections of health data.

to-fix: selected tab doesnt fill the entire area, leaving a gap between the selected area and the navtabs border, kinda annoying.